+++ src/Moon.AspNetCore.Validation/Adapters/AttributeAdapter.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc.DataAnnotations.Internal;
 using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;
 using Microsoft.Extensions.Localization;
 
-namespace Moon.AspNetCore.Validation
+namespace Moon.AspNetCore.Validation.Adapters
 {
     /// <summary>
     /// Provides a model validator for the specified validation attribute.