+++ src/Moon.AspNetCore.Validation/Adapters/MinAttributeAdapter.cs
@@ -3,7 +3,7 @@ using Microsoft.AspNetCore.Mvc.ModelBinding.Validation;
 using Microsoft.Extensions.Localization;
 using Moon.Validation;
 
-namespace Moon.AspNetCore.Validation
+namespace Moon.AspNetCore.Validation.Adapters
 {
     /// <summary>
     /// Adapter for <see cref="MinAttribute" /> attribute.