+++ src/Moon.AspNetCore.Validation/Adapters/RequiredIfAttributeAdapter.cs
@@ -4,7 +4,7 @@ using System.Globalization;
 using Microsoft.Extensions.Localization;
 using Moon.Validation;
 
-namespace Moon.AspNetCore.Validation
+namespace Moon.AspNetCore.Validation.Adapters
 {
     /// <summary>
     /// Provides a model validator for the specified validator type.