+++ src/Moon.AspNetCore.Validation/MoonValidationAttributeAdapterProvider.cs
@@ -4,6 +4,7 @@ using System.ComponentModel.DataAnnotations;
 using Microsoft.AspNetCore.Mvc.DataAnnotations;
 using Microsoft.AspNetCore.Mvc.DataAnnotations.Internal;
 using Microsoft.Extensions.Localization;
+using Moon.AspNetCore.Validation.Adapters;
 using Moon.Validation;
 
 namespace Moon.AspNetCore.Validation