+++ src/Moon.Validation/Attributes/RequiredEmptyAttribute.cs
@@ -6,24 +6,16 @@ using System.ComponentModel.DataAnnotations;
 namespace Moon.Validation
 {
     /// <summary>
-    /// Specifies that a data field value should contain only digits.
+    /// Specifies that a data field value is valid when it is null or empty.
     /// </summary>
     [AttributeUsage(AttributeTargets.Property)]
-    public class DigitsAttribute : TypeAttribute
+    public class RequiredEmptyAttribute : ValAttribute
     {
-        /// <summary>
-        /// Initializes a new instance of the <see cref="DigitsAttribute" /> class.
-        /// </summary>
-        public DigitsAttribute()
-            : base("Digits")
-        {
-        }
-
         /// <summary>
         /// Gets the default error message.
         /// </summary>
         public override string DefaultErrorMessage
-            => "The field {0} should contain only digits.";
+            => "The field {0} should be empty.";
 
         /// <summary>
         /// Determines whether the specified value is valid.
@@ -31,15 +23,6 @@ namespace Moon.Validation
         /// <param name="value">The value to validate.</param>
         /// <param name="validationContext">The validation context.</param>
         protected override bool IsValidValue(object value, ValidationContext validationContext)
-        {
-            if (value == null)
-            {
-                return true;
-            }
-
-            long result;
-            var success = long.TryParse(Convert.ToString(value), out result);
-            return success && result >= 0;
-        }
+            => string.IsNullOrEmpty(value as string);
     }
 }