+++ src/Moon.Validation/DataValidator.cs
@@ -1,4 +1,5 @@
 ﻿using System;
+using System.Collections;
 using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.Linq;
@@ -49,56 +50,86 @@ namespace Moon.Validation
         /// Returns whether the given object is valid or not.
         /// </summary>
         /// <param name="instance">The object to test. It cannot be null.</param>
-        /// <param name="validateRecursively">A value indicating whether to validate all nested objects recursively.</param>
-        public bool IsValid(object instance, bool validateRecursively = false)
-            => !Validate(instance, validateRecursively).Any();
+        /// <param name="validateNested">A value indicating whether to validate all nested objects recursively.</param>
+        public bool IsValid(object instance, bool validateNested = false)
+            => !Validate(instance, validateNested).Any();
 
         /// <summary>
         /// Validates the given object and returns an enumeration of results for the failures.
         /// </summary>
         /// <param name="instance">The object to test. It cannot be null.</param>
-        /// <param name="validateRecursively">A value indicating whether to validate all nested objects recursively.</param>
-        public IEnumerable<ValidationResult> Validate(object instance, bool validateRecursively = false)
+        /// <param name="validateNested">A value indicating whether to validate all nested objects recursively.</param>
+        public IEnumerable<ValidationResult> Validate(object instance, bool validateNested = false)
         {
             Requires.NotNull(instance, nameof(instance));
 
-            return Validate(CreateObjectContext(instance), validateRecursively);
+            return Validate(CreateObjectContext(instance), validateNested, new HashSet<object>());
         }
 
-        private IEnumerable<ValidationResult> Validate(ValidationContext objectContext, bool validateRecursively)
+        private IEnumerable<ValidationResult> Validate(ValidationContext objectContext, bool validateNested, HashSet<object> alreadyValidated)
         {
+            var instance = objectContext.ObjectInstance;
+
+            if (CanBeValidated(instance, alreadyValidated))
+            {
+                alreadyValidated.Add(instance);
+
+                if (instance is IEnumerable)
+                {
+                    return ValidateEnumerable(objectContext, validateNested, alreadyValidated);
+                }
+
+                var results = new List<ValidationResult>();
+
+                // Step 1: Validate the object properties' validation attributes
+                results.AddRange(ValidateProperties(objectContext, validateNested, alreadyValidated));
+
+                if (results.Any())
+                {
+                    return results;
+                }
+
+                // Step 2: Validate the object's validation attributes
+                results.AddRange(ValidateObject(objectContext));
+
+                if (results.Any())
+                {
+                    return results;
+                }
+
+                // Step 3: Execute IValidatableObject.Validate implementation
+                var validatable = objectContext.ObjectInstance as IValidatableObject;
+
+                if (validatable != null)
+                {
+                    results.AddRange(validatable.Validate(objectContext)
+                        .Select(x => new ValidationResult(x.ErrorMessage, MergeMemberNames(objectContext, x.MemberNames)))
+                        .Where(x => x != ValidationResult.Success));
+                }
+
+                return results;
+            }
+
+            return Enumerable.Empty<ValidationResult>();
+        }
+
+        private IEnumerable<ValidationResult> ValidateEnumerable(ValidationContext objectContext, bool validateNested, HashSet<object> alreadyValidated)
+        {
+            var index = 0;
+            var enumerable = (IEnumerable)objectContext.ObjectInstance;
             var results = new List<ValidationResult>();
 
-            // Step 1: Validate the object properties' validation attributes
-            results.AddRange(ValidateProperties(objectContext, validateRecursively));
-
-            if (results.Any())
+            foreach (var item in enumerable)
             {
-                return results;
-            }
-
-            // Step 2: Validate the object's validation attributes
-            results.AddRange(ValidateObject(objectContext));
-
-            if (results.Any())
-            {
-                return results;
-            }
-
-            // Step 3: Execute IValidatableObject.Validate implementation
-            var validatable = objectContext.ObjectInstance as IValidatableObject;
-
-            if (validatable != null)
-            {
-                results.AddRange(validatable.Validate(objectContext)
-                    .Select(x => new ValidationResult(x.ErrorMessage, AppendMemberNames(objectContext, x.MemberNames)))
-                    .Where(x => x != ValidationResult.Success));
+                var itemContext = CreateEnumerableItemContext(objectContext, index, item);
+                results.AddRange(Validate(itemContext, validateNested, alreadyValidated));
+                index++;
             }
 
             return results;
         }
 
-        private IEnumerable<ValidationResult> ValidateProperties(ValidationContext objectContext, bool validateRecursively)
+        private IEnumerable<ValidationResult> ValidateProperties(ValidationContext objectContext, bool validateRecursively, HashSet<object> alreadyValidated)
         {
             var results = new List<ValidationResult>();
             var propsToValidate = GetPropertiesToValidate(objectContext, validateRecursively);
@@ -115,7 +146,7 @@ namespace Moon.Validation
                     var memberName = property.Context.MemberName;
                     var propertyObjectContext = CreatePropertyObjectContext(objectContext, memberName, property.Value);
 
-                    propertyResults.AddRange(Validate(propertyObjectContext, true));
+                    propertyResults.AddRange(Validate(propertyObjectContext, true, alreadyValidated));
                 }
 
                 results.AddRange(propertyResults);
@@ -126,18 +157,14 @@ namespace Moon.Validation
 
         private IEnumerable<ValidationResult> ValidateObject(ValidationContext objectContext)
         {
-            var results = new List<ValidationResult>();
-
             var attributes = store.GetTypeValidationAttributes(objectContext);
-            results.AddRange(GetResults(objectContext.ObjectInstance, objectContext, attributes));
-
-            return results;
+            return GetResults(objectContext.ObjectInstance, objectContext, attributes);
         }
 
         private IEnumerable<PropertyToValidate> GetPropertiesToValidate(ValidationContext objectContext, bool validateRecursively)
         {
             var properties = objectContext.ObjectType.GetRuntimeProperties()
-                .Where(p => IsPublic(p) && p.GetIndexParameters().Empty());
+                .Where(p => p.IsPublic() && p.GetIndexParameters().Empty());
 
             foreach (var property in properties)
             {
@@ -153,15 +180,14 @@ namespace Moon.Validation
             }
         }
 
-        private IEnumerable<ValidationResult> GetResults(object value, ValidationContext validationContext, IEnumerable<ValidationAttribute> attributes)
+        private IEnumerable<ValidationResult> GetResults(object value, ValidationContext context, IEnumerable<ValidationAttribute> attributes)
         {
             var results = new List<ValidationResult>();
-            ValidationResult result;
 
             // Get the required validator; if there is one and test it first, aborting on failure
             var required = attributes.OfType<RequiredAttribute>().FirstOrDefault();
 
-            if (required != null && !TryValidate(value, validationContext, required, out result))
+            if (required != null && !TryValidate(value, context, required, out var result))
             {
                 results.Add(result);
                 return results;
@@ -170,9 +196,9 @@ namespace Moon.Validation
             // Iterate through the rest of the validators, skipping the required validator
             foreach (var attribute in attributes)
             {
-                attribute.SetOtherDisplayName(name => store.GetPropertyDisplayName(CreatePropertyContext(validationContext, name)));
+                attribute.SetOtherDisplayName(name => store.GetPropertyDisplayName(CreatePropertyContext(context, name)));
 
-                if (attribute != required && !TryValidate(value, validationContext, attribute, out result))
+                if (attribute != required && !TryValidate(value, context, attribute, out result))
                 {
                     results.Add(result);
                 }
@@ -195,6 +221,14 @@ namespace Moon.Validation
             return context;
         }
 
+        private ValidationContext CreateEnumerableItemContext(ValidationContext objectContext, int itemIndex, object item)
+        {
+            var memberName = AppendItemAccessor(objectContext, itemIndex);
+            var context = new ValidationContext(item, objectContext.Items) { MemberName = memberName };
+            context.DisplayName = store.GetTypeDisplayName(context);
+            return context;
+        }
+
         private ValidationContext CreatePropertyContext(ValidationContext objectContext, string propertyName)
         {
             var memberName = AppendMemberName(objectContext, propertyName);
@@ -203,11 +237,14 @@ namespace Moon.Validation
             return context;
         }
 
-        private IEnumerable<string> AppendMemberNames(ValidationContext objectContext, IEnumerable<string> memberNames)
+        private IEnumerable<string> MergeMemberNames(ValidationContext objectContext, IEnumerable<string> memberNames)
             => memberNames.Select(n => AppendMemberName(objectContext, n));
 
         private string AppendMemberName(ValidationContext objectContext, string memberName)
             => objectContext.MemberName != null ? $"{objectContext.MemberName}.{memberName}" : memberName;
+
+        private string AppendItemAccessor(ValidationContext objectContext, int itemIndex)
+            => objectContext.MemberName != null ? $"{objectContext.MemberName}[{itemIndex}]" : $"[{itemIndex}]";
 
         private bool TryValidate(object value, ValidationContext validationContext, ValidationAttribute attribute, out ValidationResult result)
         {
@@ -222,17 +259,32 @@ namespace Moon.Validation
             return false;
         }
 
-        private bool IsPublic(PropertyInfo property)
+        private bool CanBeValidated(object instance, ICollection<object> alreadyValidated)
         {
-            return property.GetMethod != null && property.GetMethod.IsPublic ||
-                property.SetMethod != null && property.SetMethod.IsPublic;
-        }
+            var type = instance.GetType();
+            var typeInfo = type.GetTypeInfo();
 
-        private class PropertyToValidate
-        {
-            public object Value { get; set; }
+            if (alreadyValidated.Contains(instance))
+            {
+                return false;
+            }
 
-            public ValidationContext Context { get; set; }
+            if (type == typeof(object))
+            {
+                return false;
+            }
+
+            if (typeInfo.IsSimple())
+            {
+                return false;
+            }
+
+            if (typeInfo.IsGenericType)
+            {
+                return type.GetGenericTypeDefinition() != typeof(Nullable<>);
+            }
+
+            return true;
         }
     }
 }