+++ src/Moon.Validation/PropertyInfoExtensions.cs
@@ -1,0 +1,11 @@
+﻿using System.Reflection;
+
+namespace Moon.Validation
+{
+    internal static class PropertyInfoExtensions
+    {
+        public static bool IsPublic(this PropertyInfo propertyInfo)
+            => propertyInfo.GetMethod != null && propertyInfo.GetMethod.IsPublic
+                || propertyInfo.SetMethod != null && propertyInfo.SetMethod.IsPublic;
+    }
+}