+++ src/Moon.Validation/PropertyToValidate.cs
@@ -1,0 +1,11 @@
+﻿using System.ComponentModel.DataAnnotations;
+
+namespace Moon.Validation
+{
+    internal class PropertyToValidate
+    {
+        public object Value { get; set; }
+
+        public ValidationContext Context { get; set; }
+    }
+}